Add `EXTRA_*_ARGS` for wget, curl, and singlefile
